+++ core/tauri/src/manager/mod.rs
@@ -156,6 +156,7 @@ fn replace_csp_nonce(
 }
 
 /// A resolved asset.
+#[non_exhaustive]
 pub struct Asset {
   /// The asset bytes.
   pub bytes: Vec<u8>,
@@ -163,6 +164,23 @@ pub struct Asset {
   pub mime_type: String,
   /// The `Content-Security-Policy` header value.
   pub csp_header: Option<String>,
+}
+
+impl Asset {
+  /// The asset bytes.
+  pub fn bytes(&self) -> &[u8] {
+    &self.bytes
+  }
+
+  /// The asset's mime type.
+  pub fn mime_type(&self) -> &str {
+    &self.mime_type
+  }
+
+  /// The `Content-Security-Policy` header value.
+  pub fn csp_header(&self) -> Option<&str> {
+    self.csp_header.as_deref()
+  }
 }
 
 #[default_runtime(crate::Wry, wry)]
@@ -179,6 +197,8 @@ pub struct AppManager<R: Runtime> {
   pub listeners: Listeners,
   pub state: Arc<StateManager>,
   pub config: Config,
+  #[cfg(dev)]
+  pub config_parent: Option<std::path::PathBuf>,
   pub assets: Box<dyn Assets<R>>,
 
   pub app_icon: Option<Vec<u8>>,
@@ -278,6 +298,8 @@ impl<R: Runtime> AppManager<R> {
       listeners: Listeners::default(),
       state: Arc::new(state),
       config: context.config,
+      #[cfg(dev)]
+      config_parent: context.config_parent,
       assets: context.assets,
       app_icon: context.app_icon,
       package_info: context.package_info,
@@ -456,6 +478,11 @@ impl<R: Runtime> AppManager<R> {
 
   pub fn config(&self) -> &Config {
     &self.config
+  }
+
+  #[cfg(dev)]
+  pub fn config_parent(&self) -> Option<&std::path::PathBuf> {
+    self.config_parent.as_ref()
   }
 
   pub fn package_info(&self) -> &PackageInfo {