+++ tooling/cli/src/mobile/android/dev.rs
@@ -4,7 +4,7 @@
 
 use super::{
   configure_cargo, delete_codegen_vars, device_prompt, ensure_init, env, get_app, get_config,
-  inject_assets, open_and_wait, setup_dev_config, MobileTarget,
+  inject_assets, open_and_wait, MobileTarget,
 };
 use crate::{
   dev::Options as DevOptions,
@@ -63,9 +63,6 @@ pub struct Options {
   pub open: bool,
   /// Runs on the given device name
   pub device: Option<String>,
-  /// Force prompting for an IP to use to connect to the dev server on mobile.
-  #[clap(long)]
-  pub force_ip_prompt: bool,
   /// Disable the built-in dev server for static files.
   #[clap(long)]
   pub no_dev_server: bool,
@@ -87,8 +84,8 @@ impl From<Options> for DevOptions {
       no_dev_server_wait: options.no_dev_server_wait,
       no_dev_server: options.no_dev_server,
       port: options.port,
-      force_ip_prompt: options.force_ip_prompt,
       release_mode: options.release_mode,
+      host: None,
     }
   }
 }
@@ -166,7 +163,7 @@ fn run_command(options: Options, noise_level: NoiseLevel) -> Result<()> {
 #[allow(clippy::too_many_arguments)]
 fn run_dev(
   mut interface: AppInterface,
-  mut options: Options,
+  options: Options,
   mut dev_options: DevOptions,
   tauri_config: ConfigHandle,
   device: Option<Device>,
@@ -176,13 +173,7 @@ fn run_dev(
   metadata: &AndroidMetadata,
   noise_level: NoiseLevel,
 ) -> Result<()> {
-  setup_dev_config(
-    MobileTarget::Android,
-    &mut options.config,
-    options.force_ip_prompt,
-  )?;
-
-  crate::dev::setup(&interface, &mut dev_options, tauri_config.clone(), true)?;
+  crate::dev::setup(&interface, &mut dev_options, tauri_config.clone())?;
 
   let interface_options = InterfaceOptions {
     debug: !dev_options.release_mode,
@@ -229,6 +220,7 @@ fn run_dev(
     },
     |options| {
       let cli_options = CliOptions {
+        dev: true,
         features: options.features.clone(),
         args: options.args.clone(),
         noise_level,