+++ core/tauri-macros/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## \[2.0.0-rc.0]
+
+### Dependencies
+
+- Upgraded to `tauri-codegen@2.0.0-rc.0`
+- Upgraded to `tauri-utils@2.0.0-rc.0`
+
 ## \[2.0.0-beta.19]
 
 ### What's Changed